fix(ws): correctly handle server-initiated close
This avoids a hanging behavior when the server closes the WS connection,
either via a control close frame[1] or by closing the connection
prematurely without sending the control frame (deviating from the WS
protocol).

If the control close frame is not received but the connection is closed,
the error (`close 1006 (abnormal closure): unexpected EOF`) will be
returned to the user's WS error handler, as it indicates an issue with
the server.

The user-specified close handler should be called in either case now.

Closes #581

[1]: https://tools.ietf.org/html/rfc6455#section-5.5.1